Create File Input Type
File Input Type has been created. It allows the developer to add a specific type of Files (e.x. image) and the class would generate its allowed extensions and pass it to HTML. It also provides an ability to the developer to restrict some extensions, or add some extension. This Commit contains also fixing some Typos.